+++ ListenBrainzFreshReleases/Plugin.pm
@@ -64,6 +64,7 @@ $prefs->init({
     svc_priority_qobuz    => 1,
     svc_priority_bandcamp => 2,
     svc_priority_tidal    => 3,
+    svc_priority_deezer   => 4,
 
     # Don't Stop The Music propagators (Similar / Raw / Top). dstm_count = how many
     # recommended recordings to pull from ListenBrainz into the pool; dstm_batch =